Rename and move connect web3-related hooks
