+++ src/components/Widget.tsx
@@ -2,7 +2,7 @@ import { JsonRpcProvider } from '@ethersproject/providers'
 import { TokenInfo } from '@uniswap/token-lists'
 import { Provider as Eip1193Provider } from '@web3-react/types'
 import { DEFAULT_LOCALE, SUPPORTED_LOCALES, SupportedLocale } from 'constants/locales'
-import { ActiveWeb3Provider } from 'hooks/connectWeb3/useActiveWeb3React'
+import { ActiveWeb3Provider, useActiveProvider } from 'hooks/connectWeb3/useActiveWeb3React'
 import { TransactionsUpdater } from 'hooks/transactions'
 import { BlockNumberProvider } from 'hooks/useBlockNumber'
 import { TokenListProvider } from 'hooks/useTokenList'
@@ -15,7 +15,6 @@ import styled, { keyframes } from 'styled-components/macro'
 import { Theme, ThemeProvider } from 'theme'
 import { UNMOUNTING } from 'utils/animations'
 
-import { useActiveProvider } from '../hooks/connectWeb3/useProvider'
 import { Modal, Provider as DialogProvider } from './Dialog'
 import ErrorBoundary, { ErrorHandler } from './Error/ErrorBoundary'
