+++ src/hooks/connectWeb3/useActiveWeb3React.tsx
@@ -1,11 +1,13 @@
 import { ExternalProvider, JsonRpcProvider, Web3Provider } from '@ethersproject/providers'
-import { initializeConnector, Web3ReactHooks } from '@web3-react/core'
+import { getPriorityConnector, initializeConnector, Web3ReactHooks } from '@web3-react/core'
 import { EIP1193 } from '@web3-react/eip1193'
 import { EMPTY } from '@web3-react/empty'
 import { Connector, Provider as Eip1193Provider } from '@web3-react/types'
 import { Url } from '@web3-react/url'
 import { createContext, PropsWithChildren, useContext, useEffect, useMemo } from 'react'
 import JsonRpcConnector from 'utils/JsonRpcConnector'
+
+import { connections } from './useConnect'
 
 export type Web3ContextType = {
   connector: Connector
@@ -27,11 +29,6 @@ export const Web3Context = createContext(EMPTY_CONTEXT)
 
 export default function useActiveWeb3React() {
   return useContext(Web3Context)
-}
-
-interface ActiveWeb3ProviderProps {
-  jsonRpcEndpoint?: string | JsonRpcProvider
-  provider?: Eip1193Provider | JsonRpcProvider
 }
 
 export function getNetwork(jsonRpcEndpoint?: string | JsonRpcProvider) {
@@ -62,6 +59,17 @@ function getWallet(provider?: JsonRpcProvider | Eip1193Provider) {
     return { connector, hooks }
   }
   return EMPTY_STATE
+}
+
+export function useActiveProvider(): Web3Provider | undefined {
+  const activeWalletProvider = getPriorityConnector(...connections).usePriorityProvider() as Web3Provider
+  const { connector: network } = getNetwork() // Return network-only provider if no wallet is connected
+  return activeWalletProvider ?? network.provider
+}
+
+interface ActiveWeb3ProviderProps {
+  jsonRpcEndpoint?: string | JsonRpcProvider
+  provider?: Eip1193Provider | JsonRpcProvider
 }
 
 export function ActiveWeb3Provider({
